Expose a DomainFilter connected component
This component breaks down the domains injected though its container
from the store into country, classification and subClassification, and
then make sure only unique items are on the list of countries,
classification and subClassification. Classification, and
subClassification were supposed to be unique on the upstream api, but we
realised they were not, so we patched that in.